intel_adsp: ace: add new power state
This patch adds new substate for PM_STATE_ACTIVE.

During PM_STATE_ACTIVE system is fully powered and active. If there is
no other work to do, system enters Idle thread. On Xtensa architecture
it leads to the execution of the WAITI instruction (wait for interrupt).
When system is waiting for interrupt there are possible others power
saving optimization. In WAITI the core does not perform any work, so if it
is possible to dynamically change the clock, we can force it to switch to
a lower frequency clock without harming the performance.

The new substate represents just such behavior, i.e. entering WAITI with
the lowest clock.

Signed-off-by: Tomasz Leman <tomasz.m.leman@intel.com>