dts: adsp: ace: add substates for PM_STATE_ACTIVE
This patch adds new power substates for PM_STATE_ACTIVE to reduce power
consumption during idle. The DSP can do several things in terms of clock
management to help achieve this goal.

Substate 1: allowing clock gating.

The only difference in this state is that the core goes to WAITI without
priventing clock gating in HW register. Clock gating may be enabled most
of the time, but the application must be able to prevent it if needed.
For this purpose application can use power state lock API.

Substate 2: dynamic clock switching

Intel audio DSP has the ability to change the clock depending on the
current load. The SOF application tracks the required CPU cycles per
second for each pipeline. If pipelines are active, the clock with the
desired frequency will be selected. Between the execution of scheduled
tasks, the core will often be idle (waiting for an interrupt). During
this time, the DSP can switch to a lower clock frequency without harming
performance.

Signed-off-by: Tomasz Leman <tomasz.m.leman@intel.com>